Test graph builder
